Refactor the new version mutation into root to avoid circular dependency
