+++ src/app/static/src/app/store/versions/actions.ts
@@ -1,6 +1,7 @@
+import {RootMutation} from "../root/mutations";
 import {ActionContext, ActionTree} from "vuex";
-import {initialVersionsState, VersionsState} from "../versions/versions";
-import {emptyState, RootState} from "../../root";
+import {VersionsState} from "./versions";
+import {RootState} from "../../root";
 import {api} from "../../apiService";
 import {VersionsMutations} from "./mutations";
 
@@ -10,24 +11,14 @@ export interface VersionsActions {
 
 export const actions: ActionTree<VersionsState, RootState> & VersionsActions = {
     async createVersion(context, name) {
-        const {commit, dispatch, state} = context;
+        const {commit, state }= context;
         commit({type: VersionsMutations.SetLoading, payload: true});
         await api<VersionsMutations, VersionsMutations>(context)
-            .withSuccess(VersionsMutations.NewVersion)
             .withError(VersionsMutations.VersionError)
             .postAndReturn<String>("/version/", {name})
-            .then(() => {
+            .then((response: any) => {
                 if (!state.error) {
-                    // Reset state for new version
-                    const newRootState = {
-                        ...emptyState(),
-                        versions: {
-                            ...initialVersionsState,
-                            currentVersion: state.currentVersion,
-                            currentSnapshot: state.currentVersion!.snapshots[0]
-                        }
-                    };
-                    dispatch("load/updateStoreState", newRootState, {root: true});
+                    commit({type: RootMutation.ResetVersion, payload: response.data}, { root: true });
                 }
             });
     }
